+++ tests/integration/test_spans.py
@@ -1099,3 +1099,69 @@ def test_rate_limit_indexed_consistent(
 
     spans_consumer.assert_empty()
     outcomes_consumer.assert_empty()
+
+
+def test_rate_limit_metrics_consistent(
+    mini_sentry,
+    relay_with_processing,
+    spans_consumer,
+    metrics_consumer,
+    outcomes_consumer,
+):
+    """Rate limits for total spans (i.e. metrics) are enforced consistently after metrics extraction."""
+    relay = relay_with_processing(options=TEST_CONFIG)
+    project_id = 42
+    project_config = mini_sentry.add_full_project_config(project_id)
+    project_config["config"]["features"] = [
+        "projects:span-metrics-extraction",
+        "organizations:standalone-span-ingestion",
+    ]
+    project_config["config"]["quotas"] = [
+        {
+            "categories": ["span"],
+            "limit": 3,
+            "window": 1000,
+            "id": uuid.uuid4(),
+            "reasonCode": "total_exceeded",
+        },
+    ]
+
+    spans_consumer = spans_consumer()
+    metrics_consumer = metrics_consumer()
+    outcomes_consumer = outcomes_consumer()
+
+    start = datetime.utcnow()
+    end = start + timedelta(seconds=1)
+
+    envelope = envelope_with_spans(start, end)
+
+    def summarize_outcomes():
+        counter = Counter()
+        for outcome in outcomes_consumer.get_outcomes():
+            counter[(outcome["category"], outcome["outcome"])] += outcome["quantity"]
+        return counter
+
+    # First batch passes (we over-accept once)
+    relay.send_envelope(project_id, envelope)
+    spans = list(spans_consumer.get_spans(max_attempts=4, timeout=10))
+    assert len(spans) == 4
+    metrics = list(metrics_consumer.get_metrics())
+    assert len(metrics) > 0
+    assert all(headers == [("namespace", b"spans")] for _, headers in metrics), metrics
+
+    # Accepted outcomes for main category are logged in sentry.
+    assert summarize_outcomes() == {(16, 0): 4}  # SpanIndexed, Accepted
+
+    # Second batch is limited
+    relay.send_envelope(project_id, envelope)
+    spans = list(spans_consumer.get_spans(max_attempts=1, timeout=2))
+    assert len(spans) == 0
+    metrics = list(metrics_consumer.get_metrics())
+    assert len(metrics) == 0
+    assert summarize_outcomes() == {
+        (16, 2): 4,  # SpanIndexed, RateLimited
+        (12, 2): 4,  # Span, RateLimited
+    }
+
+    spans_consumer.assert_empty()
+    outcomes_consumer.assert_empty()